+++ example_config.lua
@@ -5,6 +5,10 @@
 return {
   'zhimin/notebook_style.nvim',
   ft = 'python',
+  build = function(plugin)
+    local install = loadfile(plugin.dir .. '/lua/notebook_style/install.lua')()
+    install.run(plugin)
+  end,
   opts = {
     -- Choose border style: 'solid', 'dashed', or 'double'
     border_style = 'solid',